requirements.txt
selenium==3.141.0
splinter==0.15.0
pytest==6.2.5
allure-pytest==2.9.43